requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
streamlit==1.28.1
pandas==2.1.3
numpy==1.24.3
confluent-kafka==2.2.0
requests==2.31.0
tensorflow==2.14.0
scikit-learn==1.3.2
joblib==1.3.2
psycopg2-binary==2.9.9
redis==5.0.1